Rename `data-science` plugin to `analytics`
Also removes enabled flag
