+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/XPackSettings.java
@@ -126,10 +126,6 @@ public class XPackSettings {
     /** Setting for enabling or disabling vectors. Defaults to true. */
     public static final Setting<Boolean> VECTORS_ENABLED = Setting.boolSetting("xpack.vectors.enabled", true, Setting.Property.NodeScope);
 
-    /** Setting for enabling or disabling data science plugin. Defaults to true. */
-    public static final Setting<Boolean> DATA_SCIENCE_ENABLED = Setting.boolSetting("xpack.data-science.enabled",
-        true, Setting.Property.NodeScope);
-
     public static final List<String> DEFAULT_SUPPORTED_PROTOCOLS;
 
     static {
@@ -262,7 +258,6 @@ public class XPackSettings {
         settings.add(DATA_FRAME_ENABLED);
         settings.add(FLATTENED_ENABLED);
         settings.add(VECTORS_ENABLED);
-        settings.add(DATA_SCIENCE_ENABLED);
         return Collections.unmodifiableList(settings);
     }
