--- x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/datascience/action/DataScienceStatsAction.java
@@ -1,142 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-package org.elasticsearch.xpack.core.datascience.action;
-
-import org.elasticsearch.action.ActionType;
-import org.elasticsearch.action.FailedNodeException;
-import org.elasticsearch.action.support.nodes.BaseNodeRequest;
-import org.elasticsearch.action.support.nodes.BaseNodeResponse;
-import org.elasticsearch.action.support.nodes.BaseNodesRequest;
-import org.elasticsearch.action.support.nodes.BaseNodesResponse;
-import org.elasticsearch.cluster.ClusterName;
-import org.elasticsearch.cluster.node.DiscoveryNode;
-import org.elasticsearch.common.ParseField;
-import org.elasticsearch.common.io.stream.StreamInput;
-import org.elasticsearch.common.io.stream.StreamOutput;
-import org.elasticsearch.common.io.stream.Writeable;
-import org.elasticsearch.common.xcontent.ToXContentObject;
-import org.elasticsearch.common.xcontent.XContentBuilder;
-
-import java.io.IOException;
-import java.util.List;
-import java.util.Objects;
-
-public class DataScienceStatsAction extends ActionType<DataScienceStatsAction.Response> {
-    public static final DataScienceStatsAction INSTANCE = new DataScienceStatsAction();
-    public static final String NAME = "cluster:monitor/xpack/datascience/stats";
-
-    private DataScienceStatsAction() {
-        super(NAME, Response::new);
-    }
-
-    public static class Request extends BaseNodesRequest<Request> implements ToXContentObject {
-
-        public Request() {
-            super((String[]) null);
-        }
-
-        public Request(StreamInput in) throws IOException {
-            super(in);
-        }
-
-        @Override
-        public XContentBuilder toXContent(XContentBuilder builder, Params params) throws IOException {
-            builder.startObject();
-            builder.endObject();
-            return builder;
-        }
-
-        @Override
-        public int hashCode() {
-            // Nothing to hash atm, so just use the action name
-            return Objects.hashCode(NAME);
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (obj == null) {
-                return false;
-            }
-            if (getClass() != obj.getClass()) {
-                return false;
-            }
-            return true;
-        }
-    }
-
-    public static class NodeRequest extends BaseNodeRequest {
-        public NodeRequest(StreamInput in) throws IOException {
-            super(in);
-        }
-
-        public NodeRequest(Request request) {
-
-        }
-    }
-
-    public static class Response extends BaseNodesResponse<NodeResponse> implements Writeable, ToXContentObject {
-        public Response(StreamInput in) throws IOException {
-            super(in);
-        }
-
-        public Response(ClusterName clusterName, List<NodeResponse> nodes, List<FailedNodeException> failures) {
-            super(clusterName, nodes, failures);
-        }
-
-        @Override
-        protected List<NodeResponse> readNodesFrom(StreamInput in) throws IOException {
-            return in.readList(NodeResponse::new);
-        }
-
-        @Override
-        protected void writeNodesTo(StreamOutput out, List<NodeResponse> nodes) throws IOException {
-            out.writeList(nodes);
-        }
-
-        @Override
-        public XContentBuilder toXContent(XContentBuilder builder, Params params) throws IOException {
-            builder.startArray("stats");
-            for (NodeResponse node : getNodes()) {
-                node.toXContent(builder, params);
-            }
-            builder.endArray();
-
-            return builder;
-        }
-    }
-
-    public static class NodeResponse extends BaseNodeResponse implements ToXContentObject {
-        static ParseField CUMULATIVE_CARDINALITY_USAGE = new ParseField("cumulative_cardinality_usage");
-        private long cumulativeCardinalityUsage;
-
-        public NodeResponse(StreamInput in) throws IOException {
-            super(in);
-            cumulativeCardinalityUsage = in.readZLong();
-        }
-
-        public NodeResponse(DiscoveryNode node) {
-            super(node);
-        }
-
-        public void setCumulativeCardinalityUsage(long cumulativeCardinalityUsage) {
-            this.cumulativeCardinalityUsage = cumulativeCardinalityUsage;
-        }
-
-        @Override
-        public void writeTo(StreamOutput out) throws IOException {
-            super.writeTo(out);
-            out.writeZLong(cumulativeCardinalityUsage);
-        }
-
-        @Override
-        public XContentBuilder toXContent(XContentBuilder builder, Params params) throws IOException {
-            builder.startObject();
-            builder.field(CUMULATIVE_CARDINALITY_USAGE.getPreferredName(), cumulativeCardinalityUsage);
-            builder.endObject();
-            return builder;
-        }
-    }
-}
